# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RunRequest": {
        "additionalProperties": false,
        "properties": {
          "mode": {
            "default": "intake",
            "enum": [
              "intake",
              "comparison"
            ],
            "title": "Mode",
            "type": "string"
          },
          "revision": {
            "minimum": 1.0,
            "title": "Revision",
            "type": "integer"
          },
          "useModel": {
            "default": false,
            "title": "Usemodel",
            "type": "boolean"
          }
        },
        "required": [
          "revision"
        ],
        "title": "RunRequest",
        "type": "object"
      },
...
      "SheetInput": {
        "additionalProperties": false,
        "properties": {
          "conclusion": {
            "default": "pending",
            "enum": [
              "pending",
              "eligible",
              "eligible_after_revision",
              "ineligible"
            ],
            "title": "Conclusion",
            "type": "string"
          },
          "planningBasis": {
            "default": "",
            "enum": [
              "",
              "framework",
              "detailed",
              "sector",
              "forest"
            ],
            "title": "Planningbasis",
            "type": "string"
          },
          "recommendations": {
            "default": "",
            "maxLength": 4000,
            "title": "Recommendations",
            "type": "string"
          },
          "revision": {
            "minimum": 1.0,
            "title": "Revision",
            "type": "integer"
          },
          "sections": {
            "items": {
              "$ref": "#/components/schemas/SectionInput"
            },
            "maxItems": 5,
            "title": "Sections",
            "type": "array"
          }
        },
        "required": [
          "revision",
          "sections"
        ],
        "title": "SheetInput",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "BuildAppraisal worker",
    "version": "2026.09"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/cases/{id}/analysis": {
      "post": {
        "operationId": "start_analysis_v1_cases__id__analysis_post",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-internal-token",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Internal-Token",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Analysis"
      }
    },
    "/v1/cases/{id}/appraisal-sheet": {
      "get": {
        "operationId": "appraisal_sheet_v1_cases__id__appraisal_sheet_get",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-internal-token",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Internal-Token",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Appraisal Sheet"
      },
      "post": {
        "operationId": "save_appraisal_sheet_v1_cases__id__appraisal_sheet_post",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-internal-token",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Internal-Token",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SheetInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Save Appraisal Sheet"
      }
    },
...
    "/v1/cases/{id}/cancel": {
      "post": {
        "operationId": "cancel_v1_cases__id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-internal-token",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Internal-Token",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Mutation"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel"
      }
    },
...
    "/v1/cases/{id}/consultations": {
      "post": {
        "operationId": "consultation_v1_cases__id__consultations_post",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-internal-token",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Internal-Token",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Consultation"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Consultation"
      }
    },
...
    "/v1/cases/{id}/export/{kind}/{format}": {
      "get": {
        "operationId": "export_case_v1_cases__id__export__kind___format__get",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "kind",
            "required": true,
            "schema": {
              "enum": [
                "report",
                "supplement",
                "suspension",
                "notice",
                "decision",
                "stamp"
              ],
              "title": "Kind",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "format",
            "required": true,
            "schema": {
              "enum": [
                "pdf",
                "docx",
                "json"
              ],
              "title": "Format",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-internal-token",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Internal-Token",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export Case"
      }
    },
...
    "/v1/cases/{id}/final-review": {
      "post": {
        "operationId": "final_review_v1_cases__id__final_review_post",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-internal-token",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Internal-Token",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FinalReview"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Final Review"
      }
    },
    "/v1/cases/{id}/findings/{finding_id}": {
      "patch": {
        "operationId": "review_finding_v1_cases__id__findings__finding_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "finding_id",
            "required": true,
            "schema": {
              "title": "Finding Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-internal-token",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Internal-Token",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/app__schemas__Review"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Review Finding"
      }
    },
    "/v1/cases/{id}/internal-record/{format}": {
      "get": {
        "operationId": "internal_record_v1_cases__id__internal_record__format__get",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "format",
            "required": true,
            "schema": {
              "enum": [
                "pdf",
                "docx"
              ],
              "title": "Format",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-internal-token",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Internal-Token",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Internal Record"
      }
    },
...
    "/v1/cases/{id}/procedure-review": {
      "get": {
        "operationId": "procedure_review_v1_cases__id__procedure_review_get",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "subtype",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Subtype"
            }
          },
          {
            "in": "header",
            "name": "x-internal-token",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Internal-Token",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Procedure Review"
      },
      "post": {
        "operationId": "save_procedure_review_v1_cases__id__procedure_review_post",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-internal-token",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Internal-Token",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/app__procedure_review__Review"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Save Procedure Review"
      }
    },
    "/v1/cases/{id}/procedure-review/{kind}/{format}": {
      "get": {
        "operationId": "export_procedure_review_v1_cases__id__procedure_review__kind___format__get",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "kind",
            "required": true,
            "schema": {
              "enum": [
                "review",
                "minutes",
                "draft",
                "application"
              ],
              "title": "Kind",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "format",
            "required": true,
            "schema": {
              "enum": [
                "pdf",
                "docx"
              ],
              "title": "Format",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-internal-token",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Internal-Token",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export Procedure Review"
      }
    },
...
  }
}